feat: 회원 API 추가, 댓글 API에서 SecurityContext를 사용하도록 변경

+++ src/main/java/com/newstickr/newstickr/comment/service/CommentService.java
@@ -8,6 +8,8 @@ import com.newstickr.newstickr.news.entity.News;
 import com.newstickr.newstickr.news.repository.NewsRepository;
 import com.newstickr.newstickr.user.entity.User;
 import com.newstickr.newstickr.user.repository.UserRepository;
+import lombok.Builder;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +20,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class CommentService {
     @Autowired
     private CommentRepository commentRepository;
@@ -26,9 +29,10 @@ public class CommentService {
     @Autowired
     private UserRepository userRepository;
 
+    @Builder
     // 댓글 추가
-    public void addComment(Long newsId, CommentRequest commentRequest) {
-        Optional<User> optionalUser = userRepository.findById(commentRequest.getUserId());
+    public void addComment(Long newsId, CommentRequest commentRequest, Long userId) {
+        Optional<User> optionalUser = userRepository.findById(userId);
         if(optionalUser.isEmpty()) {
             throw new RuntimeException("User not found");
         }
@@ -48,22 +52,22 @@ public class CommentService {
             throw new RuntimeException(e);}
     }
     // 특정 댓글 조회
-    public CommentResponse getComment(Long commentId) {
-        Optional<Comment> optional = commentRepository.findById(commentId);
-        if (optional.isPresent()) {
-            Comment comment = optional.get();
-            CommentResponse commentResponse = new CommentResponse();
-            commentResponse.setCommentId(commentId);
-            commentResponse.setContent(comment.getContent());
-            commentResponse.setLikeCount(comment.getLikeCount());
-            commentResponse.setCreatedAt(comment.getCreatedAt().toString());
-            commentResponse.setUserId(comment.getUser().getId());
-            commentResponse.setProfileImg(comment.getUser().getProfileImg());
-            commentResponse.setUsername(comment.getUser().getUsername());
-            return commentResponse;
-        }
-        throw new RuntimeException("No comment with id " + commentId); // 404 처리
-    }
+//    public CommentResponse getComment(Long commentId) {
+//        Optional<Comment> optional = commentRepository.findById(commentId);
+//        if (optional.isPresent()) {
+//            Comment comment = optional.get();
+//            CommentResponse commentResponse = new CommentResponse();
+//            commentResponse.setCommentId(commentId);
+//            commentResponse.setContent(comment.getContent());
+//            commentResponse.setLikeCount(comment.getLikeCount());
+//            commentResponse.setCreatedAt(comment.getCreatedAt().toString());
+//            commentResponse.setUserId(comment.getUser().getId());
+//            commentResponse.setProfileImg(comment.getUser().getProfileImg());
+//            commentResponse.setUsername(comment.getUser().getUsername());
+//            return commentResponse;
+//        }
+//        throw new RuntimeException("No comment with id " + commentId); // 404 처리
+//    }
     // 특정 사용자가 쓴 댓글 모두 조회
     public List<CommentResponse> getAllCommentsByUserId(Long userId) {
         List<Comment> commentList =  commentRepository.findByUser_Id(userId);
@@ -94,10 +98,14 @@ public class CommentService {
     }
 
     // 댓글 수정
-    public boolean updateComment(Long commentId, CommentRequest commentRequest) {
+    public boolean updateComment(Long commentId, CommentRequest commentRequest,Long userId) {
         Optional<Comment> optional = commentRepository.findById(commentId);
         if (optional.isPresent()) {
             Comment newComment = optional.get();
+            if (newComment.getUser().getId().equals(userId)) {
+                log.info("댓글 작성자와 jwt의 userId가 달라 거부");
+                return false;
+            }
             newComment.setContent(commentRequest.getContent());
             commentRepository.save(newComment);
             return true;
@@ -105,10 +113,14 @@ public class CommentService {
         return false;
     }
     // 댓글 삭제
-    public boolean deleteComment(Long commentId) {
+    public boolean deleteComment(Long commentId, Long userId) {
         Optional<Comment> optional = commentRepository.findCommentByCommentId(commentId);
         if (optional.isPresent()) {
             Comment comment = optional.get();
+            if (comment.getUser().getId().equals(userId)) {
+                log.info("댓글 작성자와 jwt의 userId가 달라 거부");
+                return false;
+            }
             commentRepository.delete(comment);
             return true;
         }
